% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termTOgene.R
\name{termTOgene}
\alias{termTOgene}
\title{Run ivTerm using graphic interface.
The user interface contains three tabs: "Upload File", "Data Selection", and "Visualization".}
\usage{
termTOgene(shiny.maxRequestSize = 100 * 1024^2, launch.browser = T)
}
\arguments{
\item{shiny.maxRequestSize}{The maximum size of the uploaded file.}

\item{launch.browser}{If `TRUE`, launch browser.}
}
\description{
Run ivTerm using graphic interface.
The user interface contains three tabs: "Upload File", "Data Selection", and "Visualization".
}
\examples{
library(ivTerm)
termTOgene()

}
